Observe (and optional) attributes, url default
1) Observe attributes

If you want to have a Twitter button with a value like this:

socialshareText='{{ foo.bar }}'

...that value won't be changed if it's updated (by AJAX, say) after the
socialShare directive is initialized. This uses attr.$observe to be
apprised of any changes to those values.

2) Make some attributes optional

For Twitter, especially, not all attributes (like via, hashtags, etc)
are always applicable. This change doesn't add them to the string unless
they're present.

3) Make the current page the default value for socialshareUrl.